# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "description": "獲取所有上架商品列表，支援搜尋、分類篩選、價格篩選和排序",
        "operationId": "ProductController_getProducts",
        "parameters": [
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序欄位",
            "schema": {
              "enum": [
                "createdAt",
                "name"
              ],
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "最高價格",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "最低價格",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "分類ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "搜尋關鍵字（商品名稱或描述）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量，預設為10",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼，預設為1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取商品列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "獲取商品列表成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "price": {
                                "type": "number"
                              },
                              "category": {
                                "type": "object"
                              },
                              "seller": {
                                "type": "object"
                              },
                              "images": {
                                "type": "array"
                              }
                            }
                          }
                        },
                        "meta": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number"
                            },
                            "pageSize": {
                              "type": "number"
                            },
                            "total": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "獲取商品列表",
        "tags": [
          "公開商品瀏覽"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "description": "獲取指定商品的詳細資訊，包含評論和平均評分",
        "operationId": "ProductController_getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取商品詳情成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "獲取商品詳情成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        },
                        "avgRating": {
                          "type": "number"
                        },
                        "category": {
                          "type": "object"
                        },
                        "seller": {
                          "type": "object"
                        },
                        "images": {
                          "type": "array"
                        },
                        "reviews": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "summary": "獲取商品詳情",
        "tags": [
          "公開商品瀏覽"
        ]
      }
    },
...
    "/api/v1/seller/products": {
      "get": {
        "description": "獲取當前賣家的所有商品列表",
        "operationId": "ProductController_getProducts",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取商品列表成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取賣家商品列表",
        "tags": [
          "賣家商品管理"
        ]
      },
      "post": {
        "description": "賣家創建新商品",
        "operationId": "ProductController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "商品創建成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "創建商品",
        "tags": [
          "賣家商品管理"
        ]
      }
    },
    "/api/v1/seller/products/{id}": {
      "get": {
        "description": "獲取指定商品的詳細資訊",
        "operationId": "ProductController_getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取商品詳情成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取商品詳情",
        "tags": [
          "賣家商品管理"
        ]
      },
      "put": {
        "description": "更新指定商品的資訊",
        "operationId": "ProductController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "商品更新成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新商品",
        "tags": [
          "賣家商品管理"
        ]
      },
      "delete": {
        "description": "刪除指定的商品",
        "operationId": "ProductController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "商品刪除成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "刪除商品",
        "tags": [
          "賣家商品管理"
        ]
      }
    },
    "/api/v1/seller/orders": {
      "get": {
        "description": "獲取賣家相關的所有訂單",
        "operationId": "ProductController_getOrders",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取訂單列表成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取訂單列表",
        "tags": [
          "賣家商品管理"
        ]
      }
    },
    "/api/v1/seller/orders/{orderId}/ship": {
      "patch": {
        "description": "將訂單狀態更新為已出貨",
        "operationId": "ProductController_shipOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "訂單ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "訂單已出貨"
          },
          "400": {
            "description": "訂單狀態不允許出貨"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "訂單不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "訂單出貨",
        "tags": [
          "賣家商品管理"
        ]
      }
    },
    "/api/v1/seller/orders/{orderId}/complete": {
      "patch": {
        "description": "將訂單狀態更新為已完成",
        "operationId": "ProductController_completeOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "訂單ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "訂單已完成"
          },
          "400": {
            "description": "訂單狀態不允許完成"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "訂單不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "完成訂單",
        "tags": [
          "賣家商品管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zipperoo 電商平台 API",
    "description": "Zipperoo 電商平台後端 API 文檔",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}